+++ odoo/local-src/sf_account_asset_management/models/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+# Part of sensefly.
+from . import account_asset
+from . import account_asset_line