+++ odoo/local-src/sf_account_asset_management/wizard/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+# Part of sensefly.
+from . import wizard_asset_compute